Fix IE bug: call delegate with context
The Z admin logo does not show up in IE 11 due to a JavaScript error: "Invalid calling object".
The error is caused by the incorrect calling of the event attacher. The `document` context is
missing, so the browser can't figure out what to attach. The solution is to use the standard
function `call` and supply the document context.