Fix issue with an Exception being thrown when the entity being transformed had no id or value
The Exception that was thrown was the "__toString() must not throw an exception". This Exception couldn't be try-catched, after rewriting it to the $value->__toString() it was catchable.
This was caused by having an object (from a serialized value) that had an id but no other values, the (string)$value was causing the issue.